Lots of fixes Break down Tx classes Fix ZC detection Implement ZC pushing Fix spends in GUI
