+++ cppForSwig/BitcoinArmory_CppTests/BitcoinArmory_CppTests.vcxproj.filters
@@ -90,6 +90,9 @@
     <ClInclude Include="..\DatabaseBuilder.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\TxClasses.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\BinaryData.cpp">
